删除一些函数告警增加函数头文件引用。

+++ src/drivers/bsp/src/electric_energy_meter.c
@@ -253,9 +253,9 @@ void electricity_meter_analysis(uint8_t add)
     }
 
 }
-float Get_Electricity_meter_massage_energy(uint8_t add)
+double Get_Electricity_meter_massage_energy(uint8_t add)
 {
-    float electricity_data;
+    double electricity_data;
 Get_electricity_meter_massage(add,read,electric_energy_h,1);
 Get_electricity_meter_massage(add,read,electric_energy_l,1);
 electricity_data=
@@ -263,30 +263,30 @@ Electricity_meter[add].massage.massage_electric_energy_h*65536
 +Electricity_meter[add].massage.massage_electric_energy_l/100;
 return electricity_data;
 }
-float Get_Electricity_meter_massage_current(uint8_t add)
+double Get_Electricity_meter_massage_current(uint8_t add)
 {
-    float electricity_data;
+    double electricity_data;
 Get_electricity_meter_massage(add,read,current,1);
 electricity_data=Electricity_meter[add].massage.massage_ia/100;
 return electricity_data;
 }
-float Get_Electricity_meter_massage_voltage(uint8_t add)
+double Get_Electricity_meter_massage_voltage(uint8_t add)
 {
-float electricity_data;
+double electricity_data;
 Get_electricity_meter_massage(add,read,voltage,1);
 electricity_data=Electricity_meter[add].massage.massage_va/10;
 return electricity_data;
 }
-float Get_Electricity_meter_massage_power(uint8_t add)
+double Get_Electricity_meter_massage_power(uint8_t add)
 {
-    float electricity_data;
+    double electricity_data;
 Get_electricity_meter_massage(add,read,power,1);
 electricity_data=Electricity_meter[add].massage.massage_power/1000;
 return electricity_data;
 }
-float Get_Electricity_meter_massage_frequency(uint8_t add)
+double Get_Electricity_meter_massage_frequency(uint8_t add)
 {
-    float electricity_data;
+    double electricity_data;
 Get_electricity_meter_massage(add,read,frequency,1);
 electricity_data=Electricity_meter[add].massage.massage_frequency/100;
 return electricity_data;
